Add files via upload
The admin staff can handle stock levels. Each branch has a driver who delivers items to customers, each branch has stock items or different music genres. The admin staff can track how many physical copies of CDs are available. An Album_type in MW must be one of the 
following: non-promo, promotional, on-discount. The album_category is the genre the 
album belongs to. The admin is able to add albums and track the number of CDs available.